Remove obsolete Hdf5DataSetWriter. Rename Hdf5DataSetWriterFromImgPlus -> Hdf5DataSetWriter. Provide StatusService to the Hdf5DataSetWriter. Minor reformatting.
