+++ src/main/java/org/ilastik/ilastik4ij/IlastikObjectClassificationPrediction.java
@@ -1,18 +1,18 @@
 /**
  * MIT License
- *
+ * <p>
  * Copyright (c) 2017 ilastik
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in all
  * copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -20,29 +20,29 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
- *
+ * <p>
  * Author: Carsten Haubold
  */
 package org.ilastik.ilastik4ij;
 
-import org.ilastik.ilastik4ij.util.IlastikUtilities;
+import net.imagej.Dataset;
+import net.imagej.DatasetService;
+import net.imagej.ImgPlus;
 import org.ilastik.ilastik4ij.hdf5.Hdf5DataSetReader;
-import org.ilastik.ilastik4ij.hdf5.Hdf5DataSetWriterFromImgPlus;
-import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-
+import org.ilastik.ilastik4ij.hdf5.Hdf5DataSetWriter;
+import org.ilastik.ilastik4ij.util.IlastikUtilities;
 import org.scijava.ItemIO;
+import org.scijava.app.StatusService;
 import org.scijava.command.Command;
 import org.scijava.log.LogService;
 import org.scijava.options.OptionsService;
 import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
 
-import net.imagej.Dataset;
-import net.imagej.DatasetService;
-import net.imagej.ImgPlus;
+import java.io.File;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  *
@@ -55,8 +55,11 @@ public class IlastikObjectClassificationPrediction implements Command {
     LogService log;
 
     @Parameter
+    StatusService statusService;
+
+    @Parameter
     OptionsService optionsService;
-    
+
     @Parameter
     DatasetService datasetService;
 
@@ -75,7 +78,7 @@ public class IlastikObjectClassificationPrediction implements Command {
 
     @Parameter(label = "Second Input Type", choices = {"Segmentation", "Probabilities"}, style = "radioButtonHorizontal")
     private String secondInputType = "Probabilities";
-    
+
 //    @Parameter(label = "Selected Output Type", choices = {"Class Label Image", "Feature Table"}, style = "radioButtonHorizontal")
 //    private String selectedOutputType = "Class Label Image";
 
@@ -127,12 +130,12 @@ public class IlastikObjectClassificationPrediction implements Command {
             if (secondInputType.equals("Segmentation")) {
                 compressionLevel = 9;
             }
-            
+
             log.info("Dumping raw input image to temporary file " + tempInFileName);
-            new Hdf5DataSetWriterFromImgPlus(inputRawImage.getImgPlus(), tempInFileName, "data", 0, log).write();
-            
+            new Hdf5DataSetWriter(inputRawImage.getImgPlus(), tempInFileName, "data", 0, log, statusService).write();
+
             log.info("Dumping secondary input image to temporary file " + tempProbOrSegFileName);
-            new Hdf5DataSetWriterFromImgPlus(inputProbOrSegImage.getImgPlus(), tempProbOrSegFileName, "data", compressionLevel, log).write();
+            new Hdf5DataSetWriter(inputProbOrSegImage.getImgPlus(), tempProbOrSegFileName, "data", compressionLevel, log, statusService).write();
 
             if (saveOnly) {
                 log.info("Saved files for training to " + tempInFileName + " and " + tempProbOrSegFileName
@@ -188,12 +191,6 @@ public class IlastikObjectClassificationPrediction implements Command {
         } else {
             commandLine.add("--prediction_maps=" + tempProbOrSegFilename);
         }
-        
-//        if (selectedOutputType.equals("Feature Table"))
-//        {
-//            commandLine.add("--table_filename="+tempOutFileName);
-//            commandLine.add("--table_only");
-//        }
 
         log.info("Running ilastik headless command:");
         log.info(commandLine);
